refactor(llm): hoist tool-args finalizer to a shared helper
Same three-line algorithm (empty -> {}, invalid -> drop, valid -> pass
through) lived in anthropic/model.go, bedrock/model.go, and
openaicompat/model.go. Move it to llm.FinalizeToolArgs and have each
provider call through. Also collapses the openaicompat in-place
mutation variant (finalizeToolRequest on *llm.ToolRequest) into the
same byte-level helper.

No behaviour change — same inputs produce the same outputs — just
one implementation to keep honest.
