+++ llm/tool_args_test.go
@@ -1,0 +1,57 @@
+// Copyright 2026 Redpanda Data, Inc.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package llm_test
+
+import (
+	"testing"
+
+	"github.com/stretchr/testify/assert"
+
+	"github.com/redpanda-data/ai-sdk-go/llm"
+)
+
+func TestFinalizeToolArgs(t *testing.T) {
+	t.Parallel()
+
+	cases := []struct {
+		name     string
+		in       []byte
+		want     string
+		wantOK   bool
+		wantNone bool
+	}{
+		{name: "nil coerces to object", in: nil, want: "{}", wantOK: true},
+		{name: "empty coerces to object", in: []byte(""), want: "{}", wantOK: true},
+		{name: "valid object passes through", in: []byte(`{"q":"SELECT 1"}`), want: `{"q":"SELECT 1"}`, wantOK: true},
+		{name: "empty object passes through", in: []byte(`{}`), want: `{}`, wantOK: true},
+		{name: "truncated is dropped", in: []byte(`{"q":`), wantOK: false, wantNone: true},
+		{name: "garbage is dropped", in: []byte(`not json`), wantOK: false, wantNone: true},
+	}
+
+	for _, tc := range cases {
+		t.Run(tc.name, func(t *testing.T) {
+			t.Parallel()
+
+			got, ok := llm.FinalizeToolArgs(tc.in)
+			assert.Equal(t, tc.wantOK, ok)
+
+			if tc.wantNone {
+				assert.Nil(t, got)
+			} else {
+				assert.Equal(t, tc.want, string(got))
+			}
+		})
+	}
+}
